Move artifact related utils to store.artifact module
